+++ zscaler/zia/services/urlfilteringpolicies/url_filtering_rules_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/zscaler/zscaler-sdk-go/v3/tests"
 	"github.com/zscaler/zscaler-sdk-go/v3/zscaler"
 	"github.com/zscaler/zscaler-sdk-go/v3/zscaler/zia/services/browser_isolation"
+	"github.com/zscaler/zscaler-sdk-go/v3/zscaler/zia/services/common"
 )
 
 const (
@@ -73,7 +74,7 @@ func TestURLFilteringRuleIsolation(t *testing.T) {
 		Protocols:      []string{"HTTPS_RULE", "HTTP_RULE"},
 		RequestMethods: []string{"CONNECT", "DELETE", "GET", "HEAD", "OPTIONS", "OTHER", "POST", "PUT", "TRACE"},
 		UserAgentTypes: []string{"OPERA", "FIREFOX", "MSIE", "MSEDGE", "CHROME", "SAFARI", "MSCHREDGE"},
-		CBIProfile: &CBIProfile{
+		CBIProfile: &common.CBIProfile{
 			ID:   cbiProfileList[0].ID,
 			Name: cbiProfileList[0].Name,
 			URL:  cbiProfileList[0].URL,